backend/requirements.txt
# HSC-JIT v3.7 - Product Hierarchy & RAG Dependencies
# Python 3.11+ Required

# ============ Core Server & Database ============
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
redis>=5.0.1

# ============ Data Processing ============
beautifulsoup4==4.12.3
pymupdf==1.24.14
pillow==11.0.0
thefuzz==0.22.1
httpx==0.28.1
rembg
onnxruntime

# ============ AI/ML (for future enhancements) ============
sentence-transformers==3.3.1
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
google-genai>=1.0.0

# ============ Core Python ============
pydantic==2.10.5
python-dotenv==1.0.1
jinja2==3.1.4

# ============ CLI & Logging ============
rich==13.7.0
typer==0.9.0
prometheus-client==0.24.1
psutil==5.9.8

# ============ Development & Testing ============
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
ruff==0.8.4
rembg
requests
